refactor(VerifyButton): remove Spinner as never really shown
