--- src/components/atoms/Spinner/_components.spinner-css.js
@@ -1,3 +1,0 @@
-import {html} from '@polymer/lit-element';
-export default html`<style>.buv-c-spinner__flap{fill:#fff}.buv-c-spinner__flap--inverted{fill:#000}
-</style>`;
