--- src/components/atoms/Spinner/_components.spinner.scss
@@ -1,7 +1,0 @@
-.buv-c-spinner__flap {
-  fill: #fff;
-}
-
-.buv-c-spinner__flap--inverted {
-  fill: #000;
-}
