--- src/components/atoms/Spinner/index.js
@@ -1,3 +1,0 @@
-import Spinner from './Spinner';
-
-export default Spinner;
