+++ src/components/atoms/VerifyButton/_components.verify-button-css.js
@@ -1,0 +1,3 @@
+import {html} from '@polymer/lit-element';
+export default html`<style>.buv-c-verify-button{background-color:#2ab27b;border-radius:0 2px 2px 0;color:#fff;cursor:pointer;display:block;font-size:15px;font-weight:400;padding:12px 38px;border:1px solid #2ab27b}.buv-c-verify-button__label{cursor:pointer}.buv-c-verify-button--hollow{background:#fff;color:#2ab27b;font-weight:100;text-transform:uppercase;padding:12px 30px;font-size:13px}.buv-c-verify-button.is-disabled{pointer-events:none}
+</style>`;
