+++ src/components/atoms/VerifyButton/_components.verify-button.scss
@@ -26,15 +26,6 @@ $secondaryColor: #fff;
   font-size: 13px;
 }
 
-.buv-c-verify-button.has-spinner {
-  padding: 0 39px;
-}
-
-.buv-c-verify-button--hollow.has-spinner {
-  padding: 0 32px;
-  height: 42px;
-}
-
 .buv-c-verify-button.is-disabled {
   pointer-events: none;
 }
