+++ src/components/organisms/FullCertificate/FullCertificate.js
@@ -3,7 +3,7 @@ import { unsafeHTML } from 'lit-html/lib/unsafe-html.js';
 import CSS from './_components.full-certificate-css';
 import '../../atoms/CertificateDetails/index';
 import '../../atoms/FullCertificateV1';
-import '../../molecules/VerifyButton';
+import '../../atoms/VerifyButton';
 
 function renderDisplayHTML (displayHTML) {
   return html`<section class='buv-c-full-certificate'>${unsafeHTML(displayHTML)}</section>`;
