+++ test/components/b-u-v.api-certificate-invalid_test.html
@@ -27,8 +27,7 @@
               const el = fixture('api-certificate-invalid').shadowRoot.querySelectorAll('buv-raw')[0];
               const certificateInputComponent = el.shadowRoot.querySelectorAll('buv-certificate-input')[0];
               const buttonComponent = certificateInputComponent.shadowRoot.querySelectorAll('buv-verify-button')[0];
-              const rawButtonComponent = buttonComponent.shadowRoot.querySelectorAll('buv-verify-button-raw')[0];
-              const button = rawButtonComponent.shadowRoot.querySelectorAll('button')[0];
+              const button = buttonComponent.shadowRoot.querySelectorAll('button')[0];
               // allow time to process
               await wait(200);
               expect(button.disabled).to.equal(true);
