--- test/components/molecules/VerifyButton/button_test.html
@@ -1,39 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <title>Button component test suite</title>
-
-    <script src="../../../../node_modules/@webcomponents/webcomponentsjs/webcomponents-loader.js"></script>
-    <script src="../../../../node_modules/web-component-tester/browser.js"></script>
-
-    <script type="module" src="../../../../src/components/molecules/VerifyButton/VerifyButton.js"></script>
-</head>
-<body>
-    <script src="../../helpers/wait.js"></script>
-    <test-fixture id="click">
-        <template>
-            <buv-verify-button-raw isDisabled$="${false}"></buv-verify-button-raw>
-        </template>
-    </test-fixture>
-
-    <script type="module">
-      describe('Button test suite', function() {
-        describe('when the button is clicked', function () {
-          it('should show the spinner', async function () {
-            const el = fixture('click');
-            const button = el.shadowRoot.querySelectorAll('button')[0];
-
-            button.click();
-            // allow repaint
-            await wait(0);
-
-            const spinner = button.querySelectorAll('.buv-qa-spinner');
-
-            expect(spinner.length).to.equal(1);
-          });
-        });
-      });
-    </script>
-</body>
-</html>
