+++ test/components/verification/manual-verification-modal.test.html
@@ -43,8 +43,7 @@
 
               const cardCertificateComponent = el.shadowRoot.querySelectorAll('buv-card-certificate')[0];
               const verifyButtonComponent = cardCertificateComponent.shadowRoot.querySelectorAll('buv-verify-button')[0];
-              const verifyButtonRawComponent = verifyButtonComponent.shadowRoot.querySelectorAll('buv-verify-button-raw')[0];
-              const button = verifyButtonRawComponent.shadowRoot.querySelectorAll('button')[0];
+              const button = verifyButtonComponent.shadowRoot.querySelectorAll('button')[0];
 
               button.click();
