+++ test/wct-script/getTestFiles.js
@@ -1,1 +1,1 @@
-function getTestFiles () { return ["components/b-u-v.a11y_test.html","components/b-u-v.all-actions-off_test.html","components/b-u-v.allow-download-on_test.html","components/b-u-v.allow-social-share-on_test.html","components/b-u-v.api-certificate-invalid_test.html","components/b-u-v.api-certificate-valid_test.html","components/b-u-v.display-mode_test.html","components/b-u-v.input_test.html","components/b-u-v.show-metadata-on_test.html","components/b-u-v.verify-other-certificate_test.html","components/molecules/VerifyButton/button_test.html","components/verification/auto-verification-modal.test.html","components/verification/chains/verification-process-started-bitcoin-testnet.test.html","components/verification/chains/verification-process-started-bitcoin.test.html","components/verification/chains/verification-process-started-ethereum-ropsten.test.html","components/verification/chains/verification-process-started-ethereum.test.html","components/verification/chains/verification-process-started-mocknet.test.html","components/verification/success/verification-process-success-final-step-text.test.html","components/verification/success/verification-process-success-final-step.test.html","components/verification/success/verification-process-success-no-error-state.test.html","components/verification/verification-process-failing.test.html","components/verification/verification-process-started-no-final-step.test.html"] }
+function getTestFiles () { return ["components/b-u-v.a11y_test.html","components/b-u-v.all-actions-off_test.html","components/b-u-v.allow-download-on_test.html","components/b-u-v.allow-social-share-on_test.html","components/b-u-v.api-certificate-invalid_test.html","components/b-u-v.api-certificate-valid_test.html","components/b-u-v.display-mode_test.html","components/b-u-v.input_test.html","components/b-u-v.show-metadata-on_test.html","components/b-u-v.verify-other-certificate_test.html","components/verification/auto-verification-modal.test.html","components/verification/chains/verification-process-started-bitcoin-testnet.test.html","components/verification/chains/verification-process-started-bitcoin.test.html","components/verification/chains/verification-process-started-ethereum-ropsten.test.html","components/verification/chains/verification-process-started-ethereum.test.html","components/verification/chains/verification-process-started-mocknet.test.html","components/verification/success/verification-process-success-final-step-text.test.html","components/verification/success/verification-process-success-final-step.test.html","components/verification/success/verification-process-success-no-error-state.test.html","components/verification/verification-process-failing.test.html","components/verification/verification-process-started-no-final-step.test.html"] }
